in progress physics update
