+++ examples/game_2d/_old/enemy_spawner.cpp
