--- examples/game_2d/_old/game/screen_shake_system.hpp
@@ -1,15 +1,0 @@
-#pragma once
-
-#include <entt/entt.hpp>
-
-namespace game2d {
-
-struct ScreenShakeTimer
-{
-  float time_left = 0.0f;
-};
-
-void
-update_screen_shake_system(entt::registry& registry, float dt);
-
-}
