+++ examples/game_2d/src/modules/physics/helpers.hpp
@@ -1,5 +1,8 @@
 #pragma once
 
+// other lib headers
+#include <glm/glm.hpp>
+
 // c++ lib headers
 #include <map>
 #include <vector>
@@ -16,6 +19,13 @@ struct PhysicsObject
   int w;
   int h;
 };
+
+// A representation of a line, where only 1 axis matters
+struct Line
+{
+  int point_a;
+  int point_b;
+}
 
 // A collision occurs between two entities
 struct Collision2D
