+++ examples/game_2d/src/modules/physics/system.cpp
@@ -24,14 +24,11 @@ game2d::update_physics_system(entt::registry& registry, engine::Application& app
 
   SINGLETON_PhysicsComponent& p = registry.ctx<SINGLETON_PhysicsComponent>();
 
-  // ImGui::Begin("Physics Debug");
   {
     p.collidable.clear();
+    PhysicsObject po;
     const auto& view = registry.view<const CollidableComponent, const PositionIntComponent, const SizeComponent>();
-    view.each([&registry, &p](const auto entity, const auto& col, const auto& pos, const auto& size) {
-      auto& tag_comp = registry.get<TagComponent>(entity);
-
-      PhysicsObject po;
+    view.each([&registry, &po](const auto entity, const auto& col, const auto& pos, const auto& size) {
       po.ent_id = static_cast<uint32_t>(entity);
       po.x_tl = static_cast<int>(pos.x - size.w / 2.0f);
       po.y_tl = static_cast<int>(pos.y - size.h / 2.0f);
@@ -39,7 +36,6 @@ game2d::update_physics_system(entt::registry& registry, engine::Application& app
       po.h = size.h;
 
       p.collidable.push_back(po);
-      // ImGui::Text("id:%i Name:%s Layer:%i", po.ent_id, tag_comp.tag.c_str(), po.ent_id);
     });
   }
