+++ examples/game_2d/src/modules/ui_physics/system.cpp
@@ -12,8 +12,21 @@ game2d::update_ui_physics_system(entt::registry& registry, engine::Application& 
 {
   SINGLETON_PhysicsComponent& p = registry.ctx<SINGLETON_PhysicsComponent>();
 
+  int solids = 0;
+  int actors = 0;
+  const auto& coll_view = registry.view<CollidableComponent>();
+  coll_view.each([&solids, &actors](const auto& coll) {
+    if (coll.type == PhysicsType::SOLID)
+      solids += 1;
+    else {
+      actors += 1;
+    }
+  });
+
   ImGui::Begin("Physics", NULL, ImGuiWindowFlags_NoFocusOnAppearing);
   ImGui::Text("Collisions %i", p.frame_collisions.size());
   ImGui::Text("Persistent Collisions %i", p.persistent_collisions.size());
+  ImGui::Text("Solids %i", solids);
+  ImGui::Text("Actors %i", actors);
   ImGui::End();
 };
